refactor: BaseWalletAPI in a separate file from SimpleWalletApi
